social_network/requirements.txt
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-redis==5.4.0